LRCI-2857 Avoid finding the batch test class groups when generating a JSONObject
