+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/Job.java
@@ -44,6 +44,8 @@ public interface Job {
 
 	public BuildProfile getBuildProfile();
 
+	public List<BatchTestClassGroup> getDependentBatchTestClassGroups();
+
 	public List<String> getDistNodes();
 
 	public DistType getDistType();
